Adding the code for Scabbard.

+++ src/utils/AttributeType.cpp
@@ -6,7 +6,7 @@ const std::map<const BasicType, const std::string>
     AttributeType::m_typeNames({{BasicType::Integer, "int"},
                                 {BasicType::Float, "float"},
                                 {BasicType::Long, "long"},
-                                {BasicType::LongLong, "longlong"},
+                                {BasicType::LongLong, "__uint128_t"},
                                 {BasicType::Double, "double"},
                                 {BasicType::Char, "char"},
                                 {BasicType::String, "string"}});
@@ -15,4 +15,4 @@ const std::map<const std::string, const BasicType> AttributeType::m_namesToTypes
     {"int", BasicType::Integer}, {"float", BasicType::Float},
     {"long", BasicType::Long}, {"double", BasicType::Double},
     {"char", BasicType::Char}, {"string", BasicType::String},
-    {"longlong", BasicType::LongLong}};
+    {"__uint128_t", BasicType::LongLong}};
